metadata is selected correctly based on the Preview grid or pane; keyboard shortcuts working as expected.

+++ lib/misc/keyboard_handler.dart
@@ -10,6 +10,7 @@ class KeyboardHandler {
   bool isIndividualMultiSelectionPressed = false;
   bool isBlockMultiSelectionPressed = false;
   bool hasFocus = false;
+  bool isEditing = false;
   KeyboardCallback keyboardCallback;
   WidgetRef ref;
 
@@ -33,32 +34,37 @@ class KeyboardHandler {
     bool isCtrlOrMeta = _isCtrlOrMeta(event);
 
     if (event is RawKeyDownEvent) {
-      if (isCtrlOrMeta) {
-        isIndividualMultiSelectionPressed = true;
-
-        if (event.physicalKey == PhysicalKeyboardKey.keyA) {
-          keyboardCallback.selectAll();
-        }
-
-        return KeyEventResult.handled;
-      } else if (event.isShiftPressed) {
-        isBlockMultiSelectionPressed = true;
-
-        return KeyEventResult.handled;
-      } else if (event.logicalKey == LogicalKeyboardKey.arrowLeft) {
-        keyboardCallback.left();
-
-        return KeyEventResult.handled;
-      } else if (event.logicalKey == LogicalKeyboardKey.arrowRight) {
-        keyboardCallback.right();
-
-        return KeyEventResult.handled;
-      } else if (event.logicalKey == LogicalKeyboardKey.escape) {
+      if (event.logicalKey == LogicalKeyboardKey.escape) {
         keyboardCallback.exit();
         return KeyEventResult.handled;
-      } else if (event.logicalKey == LogicalKeyboardKey.backspace || event.logicalKey == LogicalKeyboardKey.delete) {
-        keyboardCallback.delete();
-        return KeyEventResult.handled;
+      } else {
+        // If we are editing, don't stuff around with the keyboard shortcuts.
+        if (!isEditing) {
+          if (isCtrlOrMeta) {
+            isIndividualMultiSelectionPressed = true;
+
+            if (event.physicalKey == PhysicalKeyboardKey.keyA) {
+              keyboardCallback.selectAll();
+            }
+
+            return KeyEventResult.handled;
+          } else if (event.isShiftPressed) {
+            isBlockMultiSelectionPressed = true;
+
+            return KeyEventResult.handled;
+          } else if (event.logicalKey == LogicalKeyboardKey.arrowLeft) {
+            keyboardCallback.left();
+
+            return KeyEventResult.handled;
+          } else if (event.logicalKey == LogicalKeyboardKey.arrowRight) {
+            keyboardCallback.right();
+
+            return KeyEventResult.handled;
+          } else if (event.logicalKey == LogicalKeyboardKey.backspace || event.logicalKey == LogicalKeyboardKey.delete) {
+            keyboardCallback.delete();
+            return KeyEventResult.handled;
+          }
+        }
       }
     } else if (event is RawKeyUpEvent) {
       if (isCtrlOrMeta) {
@@ -83,5 +89,9 @@ class KeyboardHandler {
 
   void register() {
     RawKeyboard.instance.addListener(_handleKeyEvent);
+  }
+
+  void setEditing(bool editing) {
+    isEditing = editing;
   }
 }
